+++ R/render.R
@@ -585,6 +585,18 @@ render <- function(input,
     templates <- knitr::opts_template$get()
     on.exit(knitr::opts_template$restore(templates), add = TRUE)
 
+    # specify that htmltools::htmlPreserve should use the pandoc raw
+    # attribute (e.g. ```{=html}) rather than preservation tokens when
+    # pandoc >= v2.0. Note that this option will have the intended effect
+    # only for versions of htmltools >= 0.5.0.9003.
+    if (pandoc2.0() && packageVersion("htmltools") >= "0.5.0.9003") {
+      prev <- getOption("htmltools.preserve.raw", default = NA)
+      options(htmltools.preserve.raw = TRUE)
+      if (!is.na(prev)) {
+        on.exit(options(htmltools.preserve.raw = prev), add = TRUE)
+      }
+    }
+
     # run render on_exit (run after the knit hooks are saved so that
     # any hook restoration can take precedence)
     if (is.function(output_format$on_exit))